Fixes now working!
Unfortunately it changes the server code a bit, since getaddrinfo is
async, even for AI_NUMERIC. Plus who knows maybe someone wants to bind
to terriblehack.com regardless of what its address is. So instead of
creating a server, binding it, then doing stuff with it, you give the
address you want to bind, and it gives you a server. Currently TCP only
but it'd be easy enough to do the same thing to the UDP functions, or
just pass a flag to bind so it does SOCK_STREAM or SOCK_DGRAM
(*cough*SOCK_RAW*cough*) as is convenient. Then you don't need to
duplicate code for binding to UDP and TCP.

The changes to luv_tcp_bind and luv_tcp_connect are very important. I
may not have done them right, but they enable luv to take advantage of
IPv6 addresses. The accepted practice is to use getaddrinfo to be
protocol agnostic, so that's what I did.

I got the fixes working, in that I can run luajit examples/test-tcp.lua.
Maybe will test more later when I'm not exhausted. What a waste of a
nice day this was.