Fix incorrect behaviour on Windows related to stdio pipes being read-write
Unfortunately if we always always create the stdio pipes as R+W (instead of
only R for stdin, and only W for stdout and stderr) child processes on windows
start to behave incorrectly. For example, if you spawn the 'cat' program,
write to it and then shutdown its stdin, the correct behaviour is for the
child to receive EOF and exit. But if stdin is RW the child never exits on
Windows, and you have to kill the process to force it to exit. Changing the
stdin pipe to only `UV_CREATE_PIPE|UV_READABLE_PIPE` fixes the problem.

The current solution assumes only the standard stdio fd's (0-2) are used.
A more flexible solution may be in order, but would require an API change.