CMake: Add BUILD_STATIC_LIBS option
This adds on to the changes in https://github.com/luvit/luv/pull/459 to make it possible to build module, static, and shared libraries all at the same time if requested

- To build the module, `-DBUILD_MODULE=On`; the target name for the module is `luv`
- To build a static library, `-DBUILD_STATIC_LIBS=On`; the target name for the static lib is `libluv_a` (this matches how `libuv` names its targets; `uv` for shared and `uv_a` for static)
- To build a shared library, `-DBUILD_SHARED_LIBS=On`; the target name for the shared lib is `libluv`

These are able to be mixed/matched as needed.

This is a breaking change (as was #459) for users of CMake (e.g. luvi) who will now need to specify linking against `libluv`/`libluv_a` and turn on `BUILD_STATIC_LIBS`/`BUILD_SHARED_LIBS` before calling `add_subdirectory`