Update to Libuv 1.43.0
Only new function is uv_ip_name which is a generic function to convert a binary ip4/ip6 address to a string. We only used the protocol-specific versions of this function (uv_ip4_name/uv_ip6_name) internally, and it doesn't look like we have a use for the new function internally, so it's been left unbound and added to the list of unbound (not ruled out) functions in `bindcov.sh`.