Fix crash when async fs_scandir's initial `uv_fs_scandir` call fails
If `uv_fs_scandir` returns an error, we don't want to cleanup the req since it will be cleaned up by its __gc function. Before this change, this would lead to a segfault since req->data (which would get set to NULL) would attempt to be used in `luv_fs_gc`.