Yet another luajit.cmake unwind detection fix
Previously, CMake would only read up to the first NUL or control character (for me that meant it would only read "ELF" from tmpunwind.o).
This was making CMake think there was no `eh_frame`/`__unwind_info` string in the file even if there actually was.

Now, CMake skips binary data in tmpunwind.o and just reads all the ASCII strings from the file, which is fine for what we need to do with it.

This fixes:

    In file included from luv/deps/luajit/src/ljamalg.c:23:
    luv/deps/luajit/src/lj_err.c: In function ‘lj_err_unwind_dwarf’:
    luv/deps/luajit/src/lj_err.c:469:2: error: #error "Broken build system -- only use the provided Makefiles!"
      469 | #error "Broken build system -- only use the provided Makefiles!"
          |  ^~~~~

for me on Linux x86_64 when trying to build Luv