allow coroutine continuation
Allows passing a coroutine in place of a callback in all methods using
uv_req_t. The coroutine will be yielded and then resumed using the
synchronous argument list [not the (err, value) callbacks recieve]. I
have not yet completely verified the behaviour when passing a coroutine
that is not the running thread, however it should work.

Enabled by default, any blocking function call inside of the not-main
coroutine will be a yield + resume to allow the libuv event loop to
continue running. This behavior can be changed with the
LUV_FORCE_COROUTINE_CONTINUATION flag to cmake.

Tap has been modified to *also* run tests inside of a coroutine to
provide a testing suite for these changes.

This is a merge of 5 commits:
* add coroutine continuation support
* add changes to documentation, add LUV_FORCE_COROUTINE_CONTINUATION
* run test suite over all functions run in a coroutine
* luv_error dislikes success statuses
* fix memory leak in write when yielded

Signed-off-by: truemedian <truemedian@gmail.com>