feat: support events is zero for uv_poll
As the poll() manual says that events field may be specified as
zero, in which case the only events that can be returned in revents
are POLLHUP, POLLERR, and POLLNVAL.  This patch allows luv to support
this call method with events equal zero. When the parameters in the
lua is an empty string, it means that events is zero.

Signed-off-by: Tianjia Zhang <tianjia.zhang@linux.alibaba.com>