docs: Clarify types for mode arg in some filesystem functions
Some libuv filesystem functions expect the numeric `mode` argument to be
an octal representation of the equivalent `chmod(1)` mode. This is
potentially surprising behaviour to anyone who hasn't used either
`libuv` or `glibc` from C. The documentation for e.g. `fs_mkdir` in
libuv defers to `mkdir(2)` which defers to `mode_t(3)` which only says
"It is an integer type". Furthermore, `fs_access` uses a different
`mode` type, validated using
[luv_check_amode](https://github.com/luvit/luv/blob/7233e6dea92498a244feb51b790c1ba51e8abbff/src/fs.c#L184),
which can accept strings ("symbolic mode" in `chmod(1)`). I chose not to
mention that this one still technically accepts an integer because I
have no idea about the restrictions of said integer.