luv_cfpcall: Fix stack balancing after an uncaught error
The default error handling code for printing uncaught errors uses
luaL_tolstring(). This returns the resulting string, but also pushes it onto
the stack. After we have printed it, it is safe to pop from the stack.

Leaving it on the stack causes the stack to be unbalanced.

Fixes https://github.com/luvit/luv/issues/735 (where a test case can be found)