feat(doc): move to structured format
This change introduces the files `docs.lua` and `docgen.lua`
which are used to generate `docs.md`.

`docs.lua` use a bespoke format designed so it can emit the `docs.md`
as close to the original version.

`docgen.lua` takes `docs.lua` as an input an emits `docs.md`.

To regenerate the `docs.md`, simply run `lua docgen.lua`.

Closes #632