Fix use-after-free in `new_thread`
This is caused by some strange luaL_Buffer behavior:
- luaL_Buffer pushes things on the stack and relies on their lifetime for its memory
- luaL_buffinit pushes a placeholder on the stack and then replaces it later on, but in our usage that replacement can go wrong and something other than the placeholder can be replaced. In our case, the function being dumped is what was getting replaced.
- After lua_dump, we call lua_remove with the index of the function, but since the function has been replaced with a userdata value that the buffer is using for its memory, the userdata value is removed from the stack instead
- In certain scenarios, garbage collection can act immediately and free the removed value, meaning the buffer's memory is freed
- luaL_pushresult then tries to use the buffer, which is now a use-after-free

The fix here is to avoid removing anything in particular from the stack, and instead wait until after the luaL_pushresult to move the result to the top of the stack that we intend to return with. This ensures that luv_thread_dumped only modifies the stack by +1, regardless of any intermediate values that get pushed by lua_dump/luaL_Buffer usage.

The use-after-free has only been seen in the wild when targeting x86-linux (32-bit), but that seems like a coincidence.

Closes #751