More closely match Lua 5.4's string.dump for luv's thread dumping implementation
This is not a bug fix, but it avoids some unexpected/strange Lua stack usage with lua_dump + luaL_buffinit. Without this change, after the lua_dump call the stack could have two userdata on the stack and the function being dumped could be gone (one of the userdata could erroneously replace it). After this change, the function remains on the stack and one userdata is pushed after it.

From what I can tell, Lua will push a placeholder userdata during luaL_buffinit and then try to replace that when the buffer is used, but its possible for that replacement to go haywire and replace something other than the placeholder (in our case, it can leave the placeholder on the stack and replace the function being dumped). I'm unsure if this should be considered a Lua bug, but we can be safe from any potential ill-effects by just matching Lua and calling luaL_buffinit within the lua_dump lua_Writer function.

This commit reverts d2e235503f6cb5c86121cd70cdb5d17e368bab03